fix: modal is above all the component
